Moved service data sources to its own package
